refactor: delete data protocol
